requirements.txt
langchain
python-dotenv
faiss-cpu
pypdf
tiktoken
openai
fastapi
